refactor: decouple i18n core from Qt/PySide6 import chain
Move TStr/translate/tstr/_TRANSLATIONS from iaa/application/qt/i18n.py
to iaa/i18n.py so that tasks and scheduler can import translations
without pulling in PySide6, LogBridge.install(), and the full GUI stack.

- iaa/i18n.py: pure stdlib core (locale, dataclasses, functools, typing)
- Delete iaa/application/qt/i18n.py; all 14 call sites updated
- I18nController stays as thin adapter, now imports from iaa.i18n
- iaa/application/qt/__init__.py: PEP 562 lazy import for main(),
  preventing accidental GUI bootstrap when any qt submodule is imported
