+++ iaa/application/qt/__init__.py
@@ -1,3 +1,8 @@
-from .index import main
-
 __all__ = ['main']
+
+
+def __getattr__(name):
+    if name == 'main':
+        from .index import main as _main
+        return _main
+    raise AttributeError(f"module {__name__!r} has no attribute {name!r}")
