+++ iaa/tasks/cm.py
@@ -9,7 +9,7 @@ from . import R
 from .common import go_home
 from iaa.definitions.consts import package_name
 from iaa.context import conf as get_conf, task_reporter, server
-from iaa.application.qt.i18n import TStr
+from iaa.i18n import TStr
 
 logger = logging.getLogger(__name__)
